SDK 2 -> 3 migration script
